+++ js/randomizer.js
@@ -1,0 +1,9 @@
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+const getRandomElement = (array) => (array[getRandomInteger(0, array.length)]);
+
+export {getRandomInteger, getRandomElement};
